Move Casper into casper package and renamed to CasperChat
